introduce --livecoding for native build
Now the livecoding is available to the game developpers.

The livecoding will recursively watch modifications done in the
directory. When a file has been modified in this
directory or one of its subdirectories the lua code will be reload.

It will ignore files ending with: .swp or ~   (known as backup files)
And it will ignore files starting with . to avoid watching hidden files
and avoid reloading the code if we develop on a repository like git or
subversion.

New directories created in the main directory will also be watched.