cmake: do not set -Werror for release builds
Depending on the compiler and the compiler version, more/less warnings
can be found. It means that someone wanting to make a release build of
Drystal with a compiler finding a warning that we do not have with our
compiler will not be able to build Drystal. Hence, disabling -Werror
for release builds.