abort on OOM
There is no need to handle OOM, if there is an OOM we abort immediately,
we are not core userspace there is no need to do this. Also, handling
OOM adds a lot of error paths, which needs to be tested and adds a lot of
code. If there is an OOM we are screwed so the right thing to do is
just to log the error and exit.