cmake: check if each compiler flag is supported
Instead of checking only the compiler name we directly test if the flag
is supported before adding it. Otherwise, we should also compare the
version of the compiler for specific option (e.g. -fsanitize=undefined
is introduced in gcc 4.9).

Furthermore, it allows to use other compilers than gcc/clang that
supports a subset of these flags.