cmake: add -Wdouble-promotion warning flag
It warns when there is a float value implicitly promoted to a double.
In a 32bit architecture double are emulated in software (float in
hardware) so doing something like:

float b = 1.3;
float a = 1.5 * b;

is more expensive because 1.5 * b is computed using double because 1.5
is a double and not a float.