force float for floating point computation
Unless explicitly specified a floating point literal is a double:

float a = 0.2;
float b = 1.7 * a;

1.7 * a will be computed using double and not float, we should use 1.7f
here.

On 32bit architecture float are performed in hardware and double in
software, so using double is more expensive than double.

We should use cosf() instead of cos() (same for others) and 1.5f
instead of 1.5.