cmake: remove useless c/cxx flags settings
By default CMake will add -g on debug, and -O3 on release so this is
useless. Furthermore, we should not override what has been defined
outside the cmakelists. Hence drop this and just add the flags we want
in src/
Also, we should not be the one to set these flags but the packagers.