engine: do not use gettimeofday() to get current time
This function can return time that will jump backward and forward in
time, if someone adjusts the clock it will be wrong when we
try to measure how much time we spent (e.g. start = 1000 end = 500, end
- start => ~max value of an unsigned long).
Hence, use clock_gettime with CLOCK_MONOTONIC which is not affected by
discontinuous jumps (but still affected by NTP) so it is guaranteed to
increase linearly over time.